requirements.txt
flask>=2.2,<3
flask-cors>=3.0
pandas>=1.5
numpy>=1.24
scikit-learn>=1.2
joblib>=1.2
python-dotenv>=1.0
gunicorn>=20.1
waitress>=2.1

# Optional / dev
jupyterlab>=3.0
pytest>=7.0
black>=23.0
